refactor: change name to leasing history
